+++ rbx_dom_weak/CHANGELOG.md
@@ -1,6 +1,8 @@
 # rbx\_dom\_weak Changelog
 
 ## Unreleased Changes
+
+## 1.7.0 (2019-05-14)
 * Added conversions from `BrickColor` to `Color3` and `Color3uint8`. ([#46](https://github.com/LPGhatguy/rbx-dom/pull/46))
 * Added conversions from `Color3` to `Color3uint8` and vice-versa.
 * Changed the human readable serialization of `BinaryString` values to be base64-encoded. This makes JSON-encoded values much smaller. This changes the details of the unspecified serialization of rbx_dom_weak.